remove launcher, pong and snake
