+++ Cargo.toml
@@ -12,5 +12,4 @@ figleter = { version = "0.2.2" }
 comfy-table = "7"
 cascade = "1.0"
 fltk = { version = "1.4" }
-ratatui = { version = "0.29.0" }
 glib = "0.20"
